+++ Changelog.md
@@ -4,6 +4,7 @@ All significant changes are listed in this file.
 
 ## [1.1.12] - 2022-12-03
 
+- Add context option menu to import movie subtitle
 - Pressing Ctrl-Space will play a random movie from your entire collection
 - Actresses in movie detail auto collapses with large numbers to avoid visual overflow
 - When scanning, JavLuv will no longer report manually moved movies as duplicates